Parallel summaries (#74)
Get Features Summaries parallelised. Up to n_cores - 1 can be used. Limit is not enforced in GUI, but at execution. No limit applied if Tierpsy detects it's running in a PBS or SLURM HPC (Tierpsy's estimate of n_cores might disagree with actual cores allocated) 
file_id is not guaranteed to always increase in filename_* files, as file_id is assigned before results files are processed, and results are written on a first-finished first-written basis